replaced old RoomScreenw with Rework
